tests/wpt/web-platform-tests/html/semantics/forms/the-textarea-element/value-defaultValue-textContent.html

<!DOCTYPE HTML>
<title>textarea element select() functionality</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-textarea-value">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const textarea = document.createElement("textarea");

  assert_equals(textarea.defaultValue, "", "defaultValue is empty string when it has no content");
  assert_equals(textarea.value, "", "value is empty string when it has no content");

}, "defaultValue and value are the empty string by default");

test(() => {

  const textarea = document.createElement("textarea");

  textarea.textContent = "foo bar";
  assert_equals(textarea.defaultValue, "foo bar", "the defaultValue should reflect the textContent");
  assert_equals(textarea.value, "foo bar",
    "changing the textContent should change the raw value, and subsequently the api value");

}, "defaultValue and value are affected by setting textContent");

test(() => {

  const textarea = document.createElement("textarea");

  textarea.textContent = "foo bar";
  textarea.appendChild(document.createTextNode(" baz"));
  assert_equals(textarea.defaultValue, "foo bar baz", "the defaultValue should reflect the textContent");
  assert_equals(textarea.value, "foo bar baz",
    "changing the textContent should change the raw value, and subsequently the api value");

}, "defaultValue and value are affected by textContent in combination with appending a text node");

test(() => {

  const textarea = document.createElement("textarea");

  textarea.textContent = "foo\r\nbar\rbaz\nqux";
  assert_equals(textarea.defaultValue, "foo\r\nbar\rbaz\nqux", "the defaultValue should reflect the textContent");
  assert_equals(textarea.value, "foo\nbar\nbaz\nqux", "The value property should normalize CRLF and CR to LF");

}, "defaultValue and value treat CRLF differently");

test(() => {

  const textarea = document.createElement("textarea");

  textarea.textContent = "foo";
  textarea.value = "baz";
  assert_equals(textarea.defaultValue, "foo", "setting the value property should not affect the defaultValue");
  assert_equals(textarea.textContent, "foo", "setting the value property should not affect the textContent");
  assert_equals(textarea.value, "baz",
    "on setting, the value property must set the element's raw & api value to the new value");

  textarea.value = "foo\r\nbar\rbaz\nqux";
  assert_equals(textarea.value, "foo\nbar\nbaz\nqux", "The API value should normalize CRLF and CR to LF");

  textarea.value = null;
  assert_equals(textarea.value, "", "setting the value property to null should result in an empty string");

}, "tests for the value setter");

test(() => {

  const textarea = document.createElement("textarea");

  textarea.defaultValue = "foo\0";
  assert_equals(textarea.defaultValue, "foo\0", "defaultValue after setting defaultValue");
  assert_equals(textarea.textContent, "foo\0", "textContent after setting defaultValue");
  assert_equals(textarea.value, "foo\0", "value after setting defaultValue");

  textarea.textContent = "bar\0";
  assert_equals(textarea.defaultValue, "bar\0", "defaultValue after setting textContent");
  assert_equals(textarea.textContent, "bar\0", "textContent after setting textContent");
  assert_equals(textarea.value, "bar\0", "value after setting textContent");

  textarea.value = "baz\0";
  assert_equals(textarea.defaultValue, "bar\0", "defaultValue after setting value");
  assert_equals(textarea.textContent, "bar\0", "textContent after setting value");
  assert_equals(textarea.value, "baz\0", "value after setting value");

}, "tests for U+0000 NULL");
</script>
